implemented dynamic logo and color

+++ app/src/main/java/me/alberto/findcard/data/remote/model/Country.kt
@@ -14,7 +14,7 @@ data class _Country(
     @SerializedName("currency")
     val currency: String?,
     @SerializedName("latitude")
-    val latitude: Long?,
+    val latitude: Double?,
     @SerializedName("longitude")
-    val longitude: Long?
+    val longitude: Double?
 )
